+++ src/ipfsd-in-proc.js
@@ -57,6 +57,8 @@ class Node extends EventEmitter {
         this.opts.EXPERIMENTAL.pubsub = true
       } else if (arg === '--enable-sharding-experiment') {
         this.opts.EXPERIMENTAL.sharding = true
+      } else if (arg === '--enable-namesys-pubsub') {
+        this.opts.EXPERIMENTAL.ipnsPubsub = true
       } else if (arg.startsWith('--pass')) {
         this.opts.pass = arg.split(' ').slice(1).join(' ')
       } else {
@@ -162,7 +164,7 @@ class Node extends EventEmitter {
         if (err) { return callback(err) }
         self.clean = false
         self.initialized = true
-        return callback()
+        return callback(null, this)
       })
     })
   }